Extract Renderer from Templater
This just performs the rendering of an ERB template given some
parameters, without any knowledge of how Metalware generates these
parameters or what it does with the rendered template. Extracting this
now as I think I will need to be able to do this from elsewhere in
Metalware.
