+++ src/templater.rb
@@ -28,6 +28,7 @@ require 'exceptions'
 require 'templating/iterable_recursive_open_struct'
 require 'templating/missing_parameter_wrapper'
 require 'templating/magic_namespace'
+require 'templating/renderer'
 
 
 module Metalware
@@ -95,40 +96,7 @@ module Metalware
 
     private
 
-    def replace_erb(template, template_parameters)
-      parameters_binding = template_parameters.instance_eval {binding}
-      render_erb_template(template, parameters_binding)
-    rescue NoMethodError => e
-      # May be useful to include the name of the unset parameter in this error,
-      # however this is tricky as by the time we attempt to access a method on
-      # it the unset parameter is just `nil` as far as we can see here.
-      raise UnsetParameterAccessError,
-        "Attempted to call method `#{e.name}` of unset template parameter"
-    end
-
-    def render_erb_template(template, binding)
-      # This mode allows templates to prevent inserting a newline for a given
-      # line by ending the ERB tag on that line with `-%>`.
-      trim_mode = '-'
-
-      safe_level = 0
-      erb = ERB.new(template, safe_level, trim_mode)
-
-      begin
-        erb.result(binding)
-      rescue SyntaxError => error
-        handle_error_rendering_erb(template, error)
-      end
-    end
-
-    def handle_error_rendering_erb(template, error)
-      Output.stderr "\nRendering template failed!\n\n"
-      Output.stderr "Template:\n\n"
-      Output.stderr_indented_error_message template
-      Output.stderr "\nError message:\n\n"
-      Output.stderr_indented_error_message error.message
-      abort
-    end
+    delegate :replace_erb, to: Templating::Renderer
 
     # The merging of the raw combined config files, any additional passed
     # values, and the magic `alces` namespace; this is the config prior to
